requirements.txt
apache-libcloud==2.3.0
pycrypto==2.6.1
ipdb==0.11
logging==0.4.9.6
PyYAML==3.12
